API: Add info.set_montage
