+++ mne/io/eeglab/eeglab.py
@@ -11,11 +11,10 @@ import numpy as np
 from ..utils import _read_segments_file, _find_channels
 from ..utils import _deprecate_montage
 from ..constants import FIFF
-from ..meas_info import create_info
+from ..meas_info import create_info, DEPRECATED_PARAM
 from ..base import BaseRaw
 from ...utils import logger, verbose, warn, fill_doc, Bunch
 from ...channels import make_dig_montage, make_standard_montage
-from ...channels.channels import DEPRECATED_PARAM
 from ...epochs import BaseEpochs
 from ...event import read_events
 from ...annotations import Annotations, read_annotations
